+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## vNext
 
 - Add interconnection queue to data for SPP, NYISO, ERCOT, ISONE, PJM, MISO, and CAISO
+- Add `get_generators` and `get_loads` to NYISO
 
 ## v0.12.0 - Oct 28, 2022